2-strlen.c

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -6,11 +6,13 @@
  * Return: The length of @str.
  */
 
-int _strlen(const char *str)
+size_t _strlen(const char *str)
 {
+
 	size_t length = 0;
 
 	while(*str++)
 	length++;
 	return (length);
+
 }
